Move classes into separate files 🚚
