+++ Src/Notion.Client/Models/File/FileObjectWithName.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using JsonSubTypes;
+﻿using JsonSubTypes;
 using Newtonsoft.Json;
 
 namespace Notion.Client
@@ -14,36 +13,5 @@ namespace Notion.Client
 
         [JsonProperty("name")]
         public string Name { get; set; }
-    }
-
-    public class UploadedFileWithName : FileObjectWithName
-    {
-        public override string Type => "file";
-
-        [JsonProperty("file")]
-        public Info File { get; set; }
-
-        public class Info
-        {
-            [JsonProperty("url")]
-            public string Url { get; set; }
-
-            [JsonProperty("expiry_time")]
-            public DateTime ExpiryTime { get; set; }
-        }
-    }
-
-    public class ExternalFileWithName : FileObjectWithName
-    {
-        public override string Type => "external";
-
-        [JsonProperty("external")]
-        public Info External { get; set; }
-
-        public class Info
-        {
-            [JsonProperty("url")]
-            public string Url { get; set; }
-        }
     }
 }
